bug: Custom Item 장부에 저장 안되던 버그 수정

+++ Projects/Feature/Received/Sources/Commons/PageResponseDtoSearchLedgerResponse.swift
@@ -49,7 +49,7 @@ struct SearchLedgerResponse: Codable {
 
 struct CategoryWithCustomModel: Codable {
   /// 카테고리 아이디
-  let id: Int64
+  let id: Int
   /// 카테고리 순서
   let seq: Int
   /// 카테고리 이름
